+++ docs/topics/release-notes.md
@@ -42,7 +42,7 @@ You can determine your currently installed version using `pip freeze`:
 
 ### Master
 
-* Support for custom input and output formats for `DateField`, `DateTimeField` and `TimeField`
+* Support for custom input and output formats for `DateField`, `DateTimeField` and `TimeField`.
 * Cleanup: Request authentication is no longer lazily evaluated, instead authentication is always run, which results in more consistent, obvious behavior.  Eg. Supplying bad auth credentials will now always return an error response, even if no permissions are set on the view.
 * Bugfix for serializer data being uncacheable with pickle protocol 0.
 * Bugfixes for model field validation edge-cases.